emr_project/requirements.txt.txt
Django
djangorestframework
django-rest-framework-simplejwt
google-generativeai
python-dotenv